+++ owlbot.py
@@ -20,5 +20,7 @@ node.owlbot_main(templates_excludes=[
     '.github/PULL_REQUEST_TEMPLATE.md',
     '.github/release-please.yml',
     '.github/header-checker-lint.yaml',
-    '.eslintignore'
+    '.github/workflows/ci.yaml',
+    '.eslintignore',
+    'renovate.json'
 ])