+++ src/lease-manager.ts
@@ -15,9 +15,24 @@
  */
 
 import {EventEmitter} from 'events';
+
 import {AckError, Message, Subscriber} from './subscriber';
 import {defaultOptions} from './default-options';
 import {Duration} from './temporal';
+import {DebugMessage} from './debug';
+import {logs as baseLogs} from './logs';
+
+/**
+ * Loggers. Exported for unit tests.
+ *
+ * @private
+ */
+export const logs = {
+  callbackDelivery: baseLogs.pubsub.sublog('callback-delivery'),
+  callbackExceptions: baseLogs.pubsub.sublog('callback-exceptions'),
+  expiry: baseLogs.pubsub.sublog('expiry'),
+  subscriberFlowControl: baseLogs.pubsub.sublog('subscriber-flow-control'),
+};
 
 export interface FlowControlOptions {
   allowExcessMessages?: boolean;
@@ -36,8 +51,6 @@ export interface FlowControlOptions {
  * @property {number} [maxBytes=104857600] The desired amount of memory to
  *     allow message data to consume. (Default: 100MB) It's possible that this
  *     value will be exceeded, since messages are received in batches.
- * @property {number} [maxExtensionMinutes=60] The maximum duration (in minutes)
- *     to extend the message deadline before redelivering.
  * @property {number} [maxMessages=1000] The desired number of messages to allow
  *     in memory before pausing the message stream. Unless allowExcessMessages
  *     is set to false, it is very likely that this value will be exceeded since
@@ -91,6 +104,8 @@ export class LeaseManager extends EventEmitter {
    * Adds a message to the inventory, kicking off the deadline extender if it
    * isn't already running.
    *
+   * @fires LeaseManager#full
+   *
    * @param {Message} message The message.
    * @private
    */
@@ -106,6 +121,12 @@ export class LeaseManager extends EventEmitter {
     if (allowExcessMessages! || !wasFull) {
       this._dispense(message);
     } else {
+      if (this.pending === 0) {
+        logs.subscriberFlowControl.info(
+          'subscriber for %s is client-side flow blocked',
+          this._subscriber.name,
+        );
+      }
       this._pending.push(message);
     }
 
@@ -120,10 +141,22 @@ export class LeaseManager extends EventEmitter {
   }
   /**
    * Removes ALL messages from inventory, and returns the ones removed.
+   *
+   * @fires LeaseManager#free
+   * @fires LeaseManager#empty
+   *
    * @private
    */
   clear(): Message[] {
     const wasFull = this.isFull();
+    const wasEmpty = this.isEmpty();
+
+    if (this.pending > 0) {
+      logs.subscriberFlowControl.info(
+        'subscriber for %s is unblocking client-side flow due to clear()',
+        this._subscriber.name,
+      );
+    }
 
     this._pending = [];
     const remaining = Array.from(this._messages);
@@ -133,11 +166,15 @@ export class LeaseManager extends EventEmitter {
     if (wasFull) {
       process.nextTick(() => this.emit('free'));
     }
+    if (!wasEmpty && this.isEmpty()) {
+      process.nextTick(() => this.emit('empty'));
+    }
 
     this._cancelExtension();
 
     return remaining;
   }
+
   /**
    * Indicates if we're at or over capacity.
    *
@@ -148,11 +185,23 @@ export class LeaseManager extends EventEmitter {
     const {maxBytes, maxMessages} = this._options;
     return this.size >= maxMessages! || this.bytes >= maxBytes!;
   }
+
+  /**
+   * True if we have no messages in leasing.
+   *
+   * @returns {boolean}
+   * @private
+   */
+  isEmpty(): boolean {
+    return this._messages.size === 0;
+  }
+
   /**
    * Removes a message from the inventory. Stopping the deadline extender if no
    * messages are left over.
    *
    * @fires LeaseManager#free
+   * @fires LeaseManager#empty
    *
    * @param {Message} message The message to remove.
    * @private
@@ -173,7 +222,23 @@ export class LeaseManager extends EventEmitter {
       const index = this._pending.indexOf(message);
       this._pending.splice(index, 1);
     } else if (this.pending > 0) {
+      if (this.pending > 1) {
+        logs.subscriberFlowControl.info(
+          'subscriber for %s dispensing one blocked message',
+          this._subscriber.name,
+        );
+      } else {
+        logs.subscriberFlowControl.info(
+          'subscriber for %s fully unblocked on client-side flow control',
+          this._subscriber.name,
+        );
+      }
+
       this._dispense(this._pending.shift()!);
+    }
+
+    if (this.isEmpty()) {
+      this.emit('empty');
     }
 
     if (this.size === 0 && this._isLeasing) {
@@ -225,8 +290,27 @@ export class LeaseManager extends EventEmitter {
     if (this._subscriber.isOpen) {
       message.subSpans.flowEnd();
       process.nextTick(() => {
+        message.dispatched();
+        logs.callbackDelivery.info(
+          'message (ID %s, ackID %s) delivery to user callbacks',
+          message.id,
+          message.ackId,
+        );
         message.subSpans.processingStart(this._subscriber.name);
-        this._subscriber.emit('message', message);
+        try {
+          this._subscriber.emit('message', message);
+        } catch (e: unknown) {
+          logs.callbackExceptions.error(
+            'message (ID %s, ackID %s) caused a user callback exception: %o',
+            message.id,
+            message.ackId,
+            e,
+          );
+          this._subscriber.emit(
+            'debug',
+            new DebugMessage('error during user callback', e as Error),
+          );
+        }
       });
     }
   }
@@ -265,6 +349,11 @@ export class LeaseManager extends EventEmitter {
           message.subSpans.modAckStart(deadline, false);
         }
       } else {
+        logs.expiry.warn(
+          'message (ID %s, ackID %s) has been dropped from leasing due to a timeout',
+          message.id,
+          message.ackId,
+        );
         this.remove(message);
       }
     }