+++ test/telemetry-tracing.ts
@@ -23,6 +23,7 @@ import {exporter} from './tracing';
 import {SpanKind} from '@opentelemetry/api';
 import sinon = require('sinon');
 import {PubsubMessage} from '../src/publisher';
+import {Duration} from '../src/temporal';
 
 describe('OpenTelemetryTracer', () => {
   beforeEach(() => {
@@ -212,6 +213,7 @@ describe('OpenTelemetryTracer', () => {
         topicInfo,
         message,
         'tests',
+        'create',
       );
       assert.deepStrictEqual(topicAttrs, {
         'messaging.system': 'gcp_pubsub',
@@ -222,6 +224,7 @@ describe('OpenTelemetryTracer', () => {
         'messaging.gcp_pubsub.message.exactly_once_delivery':
           message.isExactlyOnceDelivery,
         'messaging.gcp_pubsub.message.ack_id': message.ackId,
+        'messaging.operation': 'create',
         'code.function': 'tests',
       });
 
@@ -235,10 +238,12 @@ describe('OpenTelemetryTracer', () => {
         topicInfo,
         message,
         'tests',
+        'create',
       );
       assert.deepStrictEqual(topicAttrs2, {
         'messaging.system': 'gcp_pubsub',
         'messaging.destination.name': topicInfo.topicId,
+        'messaging.operation': 'create',
         'gcp.project_id': topicInfo.projectId,
         'messaging.message.envelope.size': message.data?.length,
         'code.function': 'tests',
@@ -283,10 +288,12 @@ describe('OpenTelemetryTracer', () => {
       const firstSpan = spans.pop();
       assert.ok(firstSpan);
       assert.strictEqual(firstSpan.name, `${tests.topicInfo.topicId} create`);
+      assert.strictEqual(firstSpan.attributes['messaging.operation'], 'create');
       assert.strictEqual(
         firstSpan.attributes['messaging.destination.name'],
         tests.topicInfo.topicId,
       );
+      assert.strictEqual(firstSpan.attributes['messaging.operation'], 'create');
       assert.strictEqual(
         firstSpan.attributes['messaging.system'],
         'gcp_pubsub',
@@ -312,7 +319,6 @@ describe('OpenTelemetryTracer', () => {
       const firstSpan = spans.pop();
       assert.ok(firstSpan);
       assert.strictEqual(firstSpan.name, 'other create');
-
       assert.strictEqual(
         firstSpan.attributes['messaging.destination.name'],
         'other',
@@ -343,8 +349,16 @@ describe('OpenTelemetryTracer', () => {
 
       assert.strictEqual(childReadSpan.name, 'sub subscribe');
       assert.strictEqual(
+        childReadSpan.attributes['messaging.operation'],
+        'receive',
+      );
+      assert.strictEqual(
         childReadSpan.attributes['messaging.destination.name'],
         'sub',
+      );
+      assert.strictEqual(
+        childReadSpan.attributes['messaging.operation'],
+        'receive',
       );
       assert.strictEqual(childReadSpan.kind, SpanKind.CONSUMER);
       assert.ok(childReadSpan.parentSpanId);
@@ -379,6 +393,93 @@ describe('OpenTelemetryTracer', () => {
       );
       assert.strictEqual(publishReadSpan.links.length, 1);
       assert.strictEqual(childReadSpan.links.length, 1);
+    });
+
+    it('creates ack rpc span', () => {
+      const message: PubsubMessage = {};
+      const topicName = 'projects/test/topics/topicfoo';
+      const subName = 'subTest';
+      const producerSpan = otel.PubsubSpans.createPublisherSpan(
+        message,
+        topicName,
+        'test',
+      ) as trace.Span;
+      const span = otel.PubsubSpans.createAckRpcSpan(
+        [producerSpan],
+        subName,
+        'tests',
+      );
+      assert.ok(span);
+      producerSpan.end();
+      span.end();
+
+      const spans = exporter.getFinishedSpans();
+      assert.strictEqual(spans.length, 2);
+
+      const firstSpan = spans.pop();
+      assert.ok(firstSpan);
+      assert.strictEqual(firstSpan.kind, SpanKind.CLIENT);
+      assert.strictEqual(firstSpan.name, `${subName} ack`);
+      assert.strictEqual(
+        firstSpan.attributes['messaging.destination.name'],
+        subName,
+      );
+      assert.strictEqual(
+        firstSpan.attributes['messaging.batch.message_count'],
+        1,
+      );
+      assert.strictEqual(
+        firstSpan.attributes['messaging.system'],
+        'gcp_pubsub',
+      );
+    });
+
+    it('creates modack rpc span', () => {
+      const message: PubsubMessage = {};
+      const topicName = 'projects/test/topics/topicfoo';
+      const subName = 'subTest';
+      const producerSpan = otel.PubsubSpans.createPublisherSpan(
+        message,
+        topicName,
+        'test',
+      ) as trace.Span;
+      const span = otel.PubsubSpans.createModackRpcSpan(
+        [producerSpan],
+        subName,
+        'modack',
+        'test',
+        Duration.from({seconds: 1}),
+        true,
+      );
+      assert.ok(span);
+      producerSpan.end();
+      span.end();
+
+      const spans = exporter.getFinishedSpans();
+      assert.strictEqual(spans.length, 2);
+
+      const firstSpan = spans.pop();
+      assert.ok(firstSpan);
+      assert.strictEqual(firstSpan.kind, SpanKind.CLIENT);
+      assert.strictEqual(firstSpan.name, `${subName} modack`);
+      assert.strictEqual(
+        firstSpan.attributes['messaging.destination.name'],
+        subName,
+      );
+      assert.strictEqual(
+        firstSpan.attributes[
+          'messaging.gcp_pubsub.message.ack_deadline_seconds'
+        ],
+        1,
+      );
+      assert.strictEqual(
+        firstSpan.attributes['messaging.gcp_pubsub.is_receipt_modack'],
+        true,
+      );
+      assert.strictEqual(
+        firstSpan.attributes['messaging.system'],
+        'gcp_pubsub',
+      );
     });
   });
 });